streamlit/requirements.txt
streamlit
pyyaml
httpx